Ajusta e altera tipagens e nomeclaturas

+++ src/main/java/ecommerce/dto/ProdutoComQuantidadeDTO.java
@@ -2,5 +2,5 @@ package ecommerce.dto;
 
 public record ProdutoComQuantidadeDTO(
     Long produtoId,
-    Long quantidade
+    Integer quantidade
 ) {}
